DRILL-7827: Move the Pcapng format plugin to contrib folder
