+++ contrib/format-pcapng/src/main/resources/bootstrap-format-plugins.json
@@ -1,0 +1,31 @@
+{
+  "storage": {
+    "dfs": {
+      "type": "file",
+      "formats": {
+        "pcapng": {
+          "type": "pcapng",
+          "stat" : false
+        }
+      }
+    },
+    "cp": {
+      "type": "file",
+      "formats": {
+        "pcapng": {
+          "type": "pcapng",
+          "stat" : false
+        }
+      }
+    },
+    "s3": {
+      "type": "file",
+      "formats": {
+        "pcapng": {
+          "type": "pcapng",
+          "stat" : false
+        }
+      }
+    }
+  }
+}
