+++ contrib/pom.xml
@@ -52,6 +52,7 @@
     <module>format-spss</module>
     <module>format-xml</module>
     <module>format-image</module>
+    <module>format-pcapng</module>
     <module>storage-hive</module>
     <module>storage-mongo</module>
     <module>storage-jdbc</module>
