+++ distribution/pom.xml
@@ -384,6 +384,11 @@
         </dependency>
         <dependency>
           <groupId>org.apache.drill.contrib</groupId>
+          <artifactId>drill-format-pcapng</artifactId>
+          <version>${project.version}</version>
+        </dependency>
+        <dependency>
+          <groupId>org.apache.drill.contrib</groupId>
           <artifactId>drill-format-esri</artifactId>
           <version>${project.version}</version>
         </dependency>
