+++ distribution/src/assemble/component.xml
@@ -46,6 +46,7 @@
         <include>org.apache.drill.contrib:drill-format-esri:jar</include>
         <include>org.apache.drill.contrib:drill-format-xml:jar</include>
         <include>org.apache.drill.contrib:drill-format-image:jar</include>
+        <include>org.apache.drill.contrib:drill-format-pcapng:jar</include>
         <include>org.apache.drill.contrib:drill-format-hdf5:jar</include>
         <include>org.apache.drill.contrib:drill-format-ltsv:jar</include>
         <include>org.apache.drill.contrib:drill-format-httpd:jar</include>
