+++ exec/java-exec/src/main/resources/bootstrap-storage-plugins.json
@@ -41,9 +41,6 @@
         "pcap" : {
           "type" : "pcap"
         },
-        "pcapng" : {
-          "type" : "pcapng"
-        },
         "avro" : {
           "type" : "avro",
           "extensions" : [ "avro" ]
@@ -104,9 +101,6 @@
         "pcap" : {
           "type" : "pcap"
         },
-        "pcapng" : {
-          "type" : "pcapng"
-        },
        "avro" : {
           "type" : "avro"
         },
@@ -143,9 +137,6 @@
         },
         "pcap" : {
           "type" : "pcap"
-        },
-        "pcapng" : {
-          "type" : "pcapng"
         },
         "parquet" : {
           "type" : "parquet"
